Misc fixes to resolve api connection issues
