+++ app.js
@@ -6,7 +6,7 @@ var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
-var restaurantRouter = require('./routes/restaurant');
+var restaurantRouter = require('./routes/restaurants');
 var app = express();
 
 // view engine setup
@@ -22,7 +22,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
 
-app.use('/restaurant', restaurantRouter);
+app.use('/restaurants', restaurantRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
