+++ views/partials/results.ejs
@@ -1,4 +1,4 @@
 <div>My Search Results</div>
-<!-- <% list.forEach(restaurant => { %>
+<% listOfRestaurants.forEach((restaurant) => { %>
 <p><%= restaurant.name %></p>
-<% }) %> -->
+<% }); %>
